+++ web/apps/photos/src/services/upload/uploadHttpClient.ts
@@ -1,7 +1,7 @@
 import log from "@/base/log";
 import { apiURL, uploaderOrigin } from "@/base/origins";
-import { retryAsyncOperation } from "@/gallery/utils/retry-async";
 import { EnteFile } from "@/media/file";
+import { retryAsyncOperation } from "@/utils/promise";
 import { CustomError, handleUploadError } from "@ente/shared/error";
 import HTTPService from "@ente/shared/network/HTTPService";
 import { getToken } from "@ente/shared/storage/localStorage/helpers";
