+++ web/packages/base/http.ts
@@ -1,3 +1,4 @@
+import { retryAsyncOperation } from "@/utils/promise";
 import { clientPackageName } from "./app";
 import { ensureAuthToken } from "./local-user";
 
@@ -66,3 +67,17 @@ export const ensureOk = (res: Response) => {
  */
 export const isHTTP4xxError = (e: unknown) =>
     e instanceof HTTPError && e.res.status >= 400 && e.res.status <= 499;
+
+/**
+ * A helper function to adapt {@link retryAsyncOperation} for HTTP fetches.
+ *
+ * This will ensure that the HTTP operation returning a non-200 OK status (as
+ * matched by {@link ensureOk}) is also counted as an error when considering if
+ * a request should be retried.
+ */
+export const retryEnsuringHTTPOk = (request: () => Promise<Response>) =>
+    retryAsyncOperation(async () => {
+        const r = await request();
+        ensureOk(r);
+        return r;
+    });
